requirements.txt
selenium
paho-mqtt
pyvirtualdisplay
